Helper function to estimate outcomes in individual patients

Helper function to estimate outcomes in individual patients

Usage

estimation_helper(
  n_patients,
  arm,
  index_hrqol_arm,
  first_hrqol_arm,
  final_hrqol_arm,
  inter_patient_noise_sd,
  acceleration_hrqol,
  prop_mortality_benefitters,
  mortality_trajectory_shape,
  mortality_dampening,
  mortality_rng,
  sampling_frequency = NULL,
  n_digits = 2,
  valid_hrqol_range = c(-0.757, 1),
  verbose = FALSE
)

Arguments

n_patients

single integer, the number of patients for whom to estimate outcomes

arm

character vector of length 1

first_hrqol_arm

scalar, HRQoL at ICU discharge (at arm level)

final_hrqol_arm

scalar HRQoL at end of follow-up (at arm level)

acceleration_hrqol

scalar in [0, Inf), acceleration of HRQoL improvement compared to baseline trajectory

mortality_rng

function, a function that produces times of death (only one parameter which should be the number of values to sample)

sampling_frequency

integer (single value or named vector), span between HRQoL sampling from patients

n_digits

single integer, the number of decimal places of in the first HRQoL values of patients. More digits will yield greater precision but also cause longer run-times.

valid_hrqol_range

two-element numeric vector, the lower and upper bounds of valid HRQoL values. The default (c(-0.757, 1.0)) corresponds to the Danish EQ-5D-5L index values.

verbose

single logical, should the function give progress timestamped updates? Default: TRUE

Value

data.table with outcome results for all patients. NB! There is no trial_id indicator as this is not required in all uses of this helper function.